turn appointment date time picker into a component
