+++ src/routes/provider-confirm-appointment-time.tsx
@@ -1,8 +1,9 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Col, Container, Row } from 'react-bootstrap';
 import { Helmet } from 'react-helmet';
 
 import useAccount from '@/hooks/use-account';
+import AppointmentDateTimePicker, { getDefaultAppointmentDateTime } from '@/components/appointment-date-time-picker';
 
 export const loader = () => {
 	return null;
@@ -10,6 +11,7 @@ export const loader = () => {
 
 export const Component = () => {
 	const { institution } = useAccount();
+	const [selectedAppointmentDateTime, setSelectedAppointmentDateTime] = useState(getDefaultAppointmentDateTime);
 
 	return (
 		<>
@@ -19,7 +21,12 @@ export const Component = () => {
 
 			<Container>
 				<Row>
-					<Col>TODO</Col>
+					<Col>
+						<AppointmentDateTimePicker
+							value={selectedAppointmentDateTime}
+							onChange={setSelectedAppointmentDateTime}
+						/>
+					</Col>
 				</Row>
 			</Container>
 		</>
